+++ cmd/activate_providers.go
@@ -1,0 +1,6 @@
+package cmd
+
+func init() {
+	activateCmd.AddCommand(activateGCPProviderCmd)
+	attachPreviewNotice(activateGCPProviderCmd, "GCP")
+}